third_party/blink/web_tests/wpt_internal/view-transition/pseudo-get-computed-style.html

<!DOCTYPE html>
<html class=reftest-wait>
<title>Shared transitions: check pseudo element's display property</title>
<link rel="help" href="https://github.com/WICG/view-transitions">
<link rel="author" href="mailto:vmpstr@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  contain: paint;
  page-transition-tag: target;
}
::page-transition-image-wrapper(target) {
  position: fixed;
}
</style>

<div id=target></div>

<script>
promise_test(() => {
  return new Promise(async (resolve, reject) => {
    let transition = document.startViewTransition(() => {
      assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed", ":page-transition");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute", "container(target)");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "absolute", "wrapper(target)");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute", "outgoing(target)");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute", "container(root)");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute", "wrapper(root)");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute", "outgoing(root)");

      requestAnimationFrame(() => {
        assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed", "raf :page-transition");
        assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute", "raf container(target)");
        assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "fixed", "raf wrapper(target)");

        assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute", "raf outgoing(target)");

        assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute", "raf container(root)");
        assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute", "raf wrapper(root)");
        assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute", "raf outgoing(root)");
      });
    });
    await transition.finished;
    resolve();
  });
}, "position property of pseudo elements");

promise_test(() => {
  return new Promise(async (resolve, reject) => {
    let transition = document.startViewTransition(() => {
      assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "absolute");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute");
    });

    assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "fixed");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-incoming-image(target)").position, "absolute");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-incoming-image(root)").position, "absolute");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");

    await transition.finished;

    // With custom ua sheets not applying to non-existing pseudo, the value should be the default (not absolute)
    assert_not_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");

    resolve();
  });
}, "position property of pseudo elements with prepare api");
</script>
